refactor: 리스트 뷰 어댑터 구현

+++ app/src/main/java/woowacourse/movie/dao/MovieDao.kt
@@ -25,7 +25,7 @@ class MovieDao {
         val parsedRunningTime = runningTime.toInt().minutes
         return Pair(
             title,
-            Movie(title, posterUrl, startDateTime, endDateTime, parsedRunningTime),
+            Movie(title, posterUrl.toInt(), startDateTime, endDateTime, parsedRunningTime),
         )
     }
